refactor: graphql middleware to own package
