+++ apps/cli/src/commands/tasks/tasks.command.ts
@@ -10,7 +10,6 @@ import type { CommandModule } from 'yargs'
 import { getStageOptions } from '../../shared/command'
 import { Stage } from '../../shared/utils/stage'
 import { Tasks } from '../../shared/utils/tasks'
-import { graphqlCodegen } from '../codegen'
 
 /**
  * We require this since execCommand creates a new process and any env set before that doesn't apply
