+++ libs/backend/infra/adapter/graphql/README.md
@@ -1,0 +1,3 @@
+# backend-infra-adapter-graphql
+
+This library was generated with [Nx](https://nx.dev).
