+++ libs/backend/infra/adapter/graphql/project.json
@@ -1,0 +1,16 @@
+{
+  "name": "backend-infra-adapter-graphql",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/backend/infra/adapter/graphql/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/backend/infra/adapter/graphql/**/*.ts"]
+      }
+    }
+  },
+  "tags": ["scope:backend"]
+}
