+++ libs/backend/infra/adapter/graphql/src/cors.middleware.ts
@@ -1,0 +1,19 @@
+import type { NextApiHandler } from 'next'
+
+export const corsMiddleware: NextApiHandler = (req, res) => {
+  res.setHeader('Access-Control-Allow-Credentials', 'true')
+  res.setHeader('Access-Control-Allow-Origin', '*')
+  res.setHeader('Access-Control-Allow-Headers', '*')
+  res.setHeader('Access-Control-Allow-Methods', '*')
+
+  /**
+   * Not an actual request
+   */
+  if (req.method === 'OPTIONS') {
+    res.end()
+
+    return false
+  }
+
+  return true
+}
